improved generation
